Reshape slider knobs to circles

+++ Source/PluginEditor.h
@@ -42,7 +42,7 @@ struct RotarySliderWithLabels : juce::Slider
     }
     
     void paint(juce::Graphics& g) override;
-    juce::Rectangle<int> getSliderBounds() const;
+    juce::Rectangle<float> getSliderBounds() const;
     int getTextHeight() const { return 14; }
     juce::String getDisplayString() const;
